alsa_sound: Fix buffer pointer mismatch in proxy
-There were glitches heard at times when listening
to music over Bluetooth
-The issue was that due to slight scheduling delays
when the proxy read thread in the HAL wakes up
there could be more than one period size worth
of data in the proxy. This would result
in the proxy copying out only one period
but incrementing the buffer by whatever thats
available. This resulted in a glitch
-To fix this, we have to make sure that the
pointer is incremented only by the amount of data
thats actually read from the proxy

Change-Id: I42e689ccfad2548d1bbbbc13995fb3b649593bbd